Feat: Removed AsMu V1 -> V2
