+++ src/lib/blocks/Javascript/text.ts
@@ -6,7 +6,7 @@ import ValueInput from "$lib/utils/BlockGen/Inputs/ValueInput";
 import Placeholder from "$lib/utils/ToolboxGen/Placeholder";
 import StatementInput from "$lib/utils/BlockGen/Inputs/StatementInput";
 import Dropdown from "$lib/utils/BlockGen/Inputs/Dropdown";
-import AssemblerMutatorV2 from "$lib/utils/BlockGen/Mutators/AssemblerMutatorV2";
+import AssemblerMutatorV2 from "$lib/utils/BlockGen/Mutators/AssemblerMutator";
 
 const blocks: BlockDefinition[] = [
 	{
